Work around problem with fluent-plugin-ec2-metadata on Ubuntu.
Right now, latest version of fluent-plugin-ec2-metadata requires just any
version of the fluentd gem. The latest versions of which ( 0.14.x ) now
require ruby > 2.0, which, unfortunately, isn't compatible with fluentd
on Ubuntu.

So, the correct solution would be to follow the 0.12.x branch of the
fluentd plugin.

Submitted takus/fluent-plugin-ec2-metadata#17 upstream

In the meantime, I've made a copy of the gem with just a clearer
set of requirements to use.